hello_world/pyproject.toml
[tool.poetry]
name = "hello-world"
version = "0.1.0"
description = ""
authors = ["Vincent Groves <vincent.groves@leanix.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pyjwt = "^2.9.0"
cryptography = "^43.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
